adapter: remove catalog-based timetamp oracle and machinery that supports it
We are now using the Postgres/Crdb-backed oracle in production.
